Removed and changed Vision of the Past to fit in the file from customer
added Chest locations for achievment into the Vision of past mission. also created a folder for this mission since there are more files to be created in the future so in order to keep the stuff organised i added the folder.
also it helps for better overview cause there are other missions called "Visions of the past"